requirements.txt
openai-whisper
gTTS>=2.5.0
pydub>=0.25.1
streamlit>=1.30.0
deepl>=1.18.0
nltk>=3.8.1
googletrans==4.0.0rc1
